+++ frame-metadata/Cargo.toml
@@ -14,7 +14,8 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
 codec = { package = "parity-scale-codec", version = "1.3.4", default-features = false, features = ["derive"] }
-scale-info = { path = "../../scale-info", package = "scale-info", default-features = false, optional = true, features = ["derive"] }
+cfg-if = "1.0.0"
+scale-info = { git = "https://github.com/paritytech/scale-info", default-features = false, optional = true, features = ["derive"] }
 serde = { version = "1.0.101", optional = true, features = ["derive"] }
 
 [features]
